fix cpp07ex02

+++ cpp07/ex02/Array.hpp
@@ -7,15 +7,15 @@ template <class Elm>
 class Array {
 private:
   Elm *_data;
-  size_t _size;
+  unsigned int _size;
 
 public:
   Array() : _data(NULL), _size(0) {}
-  Array(size_t n) : _data(new Elm[n]()), _size(n) {}
+  Array(unsigned int n) : _data(new Elm[n]()), _size(n) {}
 
   Array(const Array &other)
       : _data(new Elm[other._size]()), _size(other._size) {
-    for (size_t i = 0; i < _size; i++) {
+    for (unsigned int i = 0; i < _size; i++) {
       _data[i] = other._data[i];
     }
   }
@@ -28,25 +28,25 @@ public:
     delete[] _data;
     _data = new Elm[rhs._size]();
     _size = rhs._size;
-    for (size_t i = 0; i < _size; i++) {
+    for (unsigned int i = 0; i < _size; i++) {
       _data[i] = rhs._data[i];
     }
     return *this;
   }
 
-  const Elm &operator[](size_t index) const {
+  const Elm &operator[](unsigned int index) const {
     if (index >= _size || index < 0 || _size == 0)
       throw std::out_of_range("Access to out of range!");
     return _data[index];
   }
 
-  Elm &operator[](size_t index) {
+  Elm &operator[](unsigned int index) {
     if (index >= _size || index < 0 || _size == 0)
       throw std::out_of_range("Access to out of range!");
     return _data[index];
   }
 
-  size_t size() const { return _size; }
+  unsigned int size() const { return _size; }
 };
 
 template <typename T>
